split subtitles into 2 components
