+++ force-app/main/default/lwc/nksSubtitleLinker/__tests__/nksSubtitleLinker.test.js
@@ -1,0 +1,25 @@
+import { createElement } from 'lwc';
+import NksSubtitleLinker from 'c/nksSubtitleLinker';
+
+describe('c-nks-subtitle-linker', () => {
+    afterEach(() => {
+        // The jsdom instance is shared across test cases in a single file so reset the DOM
+        while (document.body.firstChild) {
+            document.body.removeChild(document.body.firstChild);
+        }
+    });
+
+    it('TODO: test case generated by CLI command, please fill in test logic', () => {
+        // Arrange
+        const element = createElement('c-nks-subtitle-linker', {
+            is: NksSubtitleLinker
+        });
+
+        // Act
+        document.body.appendChild(element);
+
+        // Assert
+        // const div = element.shadowRoot.querySelector('div');
+        expect(1).toBe(1);
+    });
+});
