+++ force-app/main/default/lwc/nksSubtitleLinker/nksSubtitleLinker.html
@@ -1,0 +1,21 @@
+<template>
+    <div if:true={isSubtitleFile}>
+        <h1 class="slds-text-heading_small slds-text-title_bold slds-p-bottom_xx-small">Undertekstlenke</h1>
+        <div class="slds-box slds-box_x-small slds-theme_shade">
+            <lightning-layout multiple-rows>
+                <lightning-layout-item alignment-bump="right" size="11">
+                    <lightning-formatted-text value={videoTrackURL}></lightning-formatted-text>
+                </lightning-layout-item>
+                <lightning-layout-item>
+                    <lightning-button-icon
+                        data-id="subtitle-copy-button"
+                        icon-name="utility:copy"
+                        alternative-text="Kopier"
+                        onclick={handleCopy}
+                        variant="brand"
+                    ></lightning-button-icon>
+                </lightning-layout-item>
+            </lightning-layout>
+        </div>
+    </div>
+</template>
