+++ force-app/main/default/lwc/nksVideoDetails/nksVideoDetails.html
@@ -5,5 +5,6 @@
     <div class="slds-p-top_small" if:true={isVideoFile}>
         <c-nks-embed-generator record-id={recordId}></c-nks-embed-generator>
     </div>
-    <c-nks-thumbnail-generator record-id={recordId} filetype={wiredData}></c-nks-thumbnail-generator>
+    <c-nks-thumbnail-generator record-id={recordId} filetype={filetype}></c-nks-thumbnail-generator>
+    <c-nks-subtitle-linker record-id={recordId} filetype={filetype}></c-nks-subtitle-linker>
 </template>
